+++ src/day_7.rs
@@ -3,12 +3,12 @@ use std::collections::VecDeque;
 use itertools::Itertools;
 
 #[aoc_generator(day7)]
-pub fn input_generator(intcode: &str) -> Vec<isize> {
+pub fn input_generator(intcode: &str) -> Vec<i64> {
     intcode.split(',').flat_map(|n| n.parse()).collect()
 }
 
 #[aoc(day7, part1)]
-pub fn find_highest_signal(intcode: &[isize]) -> isize {
+pub fn find_highest_signal(intcode: &[i64]) -> i64 {
     let mut highest_signal = 0;
     for permutation in (0..5).permutations(5) {
         let mut next_input = 0;
@@ -29,7 +29,7 @@ pub fn find_highest_signal(intcode: &[isize]) -> isize {
 }
 
 #[aoc(day7, part2)]
-pub fn find_highest_signal_with_feedback_loop(intcode: &[isize]) -> isize {
+pub fn find_highest_signal_with_feedback_loop(intcode: &[i64]) -> i64 {
     let mut highest_signal = 0;
     for permutation in (5..10).permutations(5) {
         let mut amplifiers = [
